Only install deps on first go

+++ shell/install-dependencies.sh
@@ -1,8 +1,13 @@
 #!/bin/sh
 
-sudo rpm -Uvh http://mirror.webtatic.com/yum/el6/latest.rpm
+if [ ! -f /.initial-setup-has-run ]
+then
+  sudo rpm -Uvh http://mirror.webtatic.com/yum/el6/latest.rpm
 
-sudo gem install librarian-puppet
-sudo yum -y install git
+  sudo gem install librarian-puppet
+  sudo yum -y install git
+
+  touch /.initial-setup-has-run
+fi
 
 cd /vagrant/puppet && librarian-puppet install
